Added common types component for iam_private
